GEMM: move unit tests to Blas
