+++ batched/dense/unit_test/Test_Batched_Dense.hpp
@@ -7,9 +7,6 @@
 #include "Test_Batched_SerialAxpy_Complex.hpp"
 #include "Test_Batched_SerialEigendecomposition.hpp"
 #include "Test_Batched_SerialEigendecomposition_Real.hpp"
-#include "Test_Batched_SerialGemm.hpp"
-#include "Test_Batched_SerialGemm_Real.hpp"
-#include "Test_Batched_SerialGemm_Complex.hpp"
 #include "Test_Batched_BatchedGemm.hpp"
 #include "Test_Batched_BatchedGemm_Real.hpp"
 #include "Test_Batched_BatchedGemm_Complex.hpp"
@@ -42,9 +39,6 @@
 #include "Test_Batched_TeamAxpy.hpp"
 #include "Test_Batched_TeamAxpy_Real.hpp"
 #include "Test_Batched_TeamAxpy_Complex.hpp"
-#include "Test_Batched_TeamGemm.hpp"
-#include "Test_Batched_TeamGemm_Real.hpp"
-#include "Test_Batched_TeamGemm_Complex.hpp"
 #include "Test_Batched_TeamGesv.hpp"
 #include "Test_Batched_TeamGesv_Real.hpp"
 #include "Test_Batched_TeamInverseLU.hpp"
@@ -69,9 +63,6 @@
 #include "Test_Batched_TeamVectorAxpy_Complex.hpp"
 #include "Test_Batched_TeamVectorEigendecomposition.hpp"
 #include "Test_Batched_TeamVectorEigendecomposition_Real.hpp"
-#include "Test_Batched_TeamVectorGemm.hpp"
-#include "Test_Batched_TeamVectorGemm_Real.hpp"
-#include "Test_Batched_TeamVectorGemm_Complex.hpp"
 #include "Test_Batched_TeamVectorGesv.hpp"
 #include "Test_Batched_TeamVectorGesv_Real.hpp"
 #include "Test_Batched_TeamVectorQR.hpp"
