+++ blas/unit_test/Test_Blas.hpp
@@ -52,6 +52,9 @@
 
 // Blas 3
 #include "Test_Blas3_gemm.hpp"
+#include "Test_Blas3_serial_gemm.hpp"
+#include "Test_Blas3_team_gemm.hpp"
+#include "Test_Blas3_teamvector_gemm.hpp"
 #include "Test_Blas3_trmm.hpp"
 #include "Test_Blas3_trsm.hpp"
